Transition most to SCSS source
For some reason, despite the fact that it should be exactly equivalent, the compiled header.scss has some strange issues on Chrome that are not present with the original header.css
These issues occur in pageheaderlinks2 where it appears the text for links like "PERSONAL INFORMATION" do not update when their parents' formatting changes despite having color: unset. This results in the final color not matching any of the styles in the dropdown of Inspect Element > Element > Computed. This suggests the issue may related to Chrome's implementation of inherited styles.